+++ markdown-link-check-config.json
@@ -1,0 +1,1 @@
+{"ignorePatterns": [{"pattern": "^https://docs.google.com"}]}
