+++ tf2_web_republisher/CMakeLists.txt
@@ -19,16 +19,24 @@ foreach(package IN ITEMS ${THIS_PACKAGE_INCLUDE_DEPENDS})
   find_package(${package} REQUIRED)
 endforeach()
 
-add_executable(${PROJECT_NAME} src/tf_web_republisher.cpp)
+add_library(tf2_web_republisher src/tf2_web_republisher.cpp)
 target_include_directories(
-  ${PROJECT_NAME} PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-                         $<INSTALL_INTERFACE:include>)
+  tf2_web_republisher
+  PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+         $<INSTALL_INTERFACE:include>)
 
 ament_target_dependencies(${PROJECT_NAME} ${THIS_PACKAGE_INCLUDE_DEPENDS})
 
+add_executable(tf2_web_republisher_node src/tf2_web_republisher_node.cpp)
+target_link_libraries(tf2_web_republisher_node tf2_web_republisher)
+
+if(BUILD_TESTING)
+  add_subdirectory(test)
+endif()
+
 install(
-  TARGETS ${PROJECT_NAME}
-  EXPORT export_${PROJECT_NAME}
+  TARGETS tf2_web_republisher_node
+  EXPORT export_test_tf2_web_republisher_node
   RUNTIME DESTINATION lib/${PROJECT_NAME})
 
 install(DIRECTORY include/ DESTINATION include)
