+++ tf2_web_republisher/src/tf2_web_republisher_node.cpp
@@ -1,0 +1,13 @@
+#include <rclcpp/rclcpp.hpp>
+#include "tf_web_republisher/tf2_web_republisher.hpp"
+
+int main(int argc, char** argv)
+{
+  rclcpp::init(argc, argv);
+
+  auto tf2_web_republisher = std::make_shared<TFRepublisher>("tf2_web_republisher");
+
+  rclcpp::spin(tf2_web_republisher);
+
+  return 0;
+}
