+++ tf2_web_republisher/test/test_tf2_web_republisher.cpp
@@ -1,0 +1,13 @@
+#include <gmock/gmock.h>
+
+#include "tf_web_republisher/tf2_web_republisher.hpp"
+
+namespace
+{
+
+TEST(TFWebRepublisher, TestActionCall)
+{
+    EXPECT_EQ(true, true);
+}
+
+}  // namespace moveit_pro_controllers
